pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "i19-bluesky"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
description = "Bluesky tools for i19 beamline."
dependencies = [
    "numpy",
    "nexgen",
    "pydantic",
    "pydantic-extra-types >= 2.10.1",
    "requests",
    "fastapi[all]",
    "mysql-connector-python == 9.5.0", # see https://github.com/DiamondLightSource/mx-bluesky/issues/1568
    # These dependencies may be issued as pre-release versions and should have a pin constraint
    # as by default pip-install will not upgrade to a pre-release.
    # Blueapi pinned until GDA can be updated on the beamline to use latest version
    "blueapi >= 1.4.1",
    "ophyd ==1.11.0",
    "ophyd-async[ca,pva] >= 0.15.0",
    "bluesky >= 1.13",
    "dls-dodal @ git+https://github.com/DiamondLightSource/dodal.git@main",
]
dynamic = ["version"]
license.file = "LICENSE"
readme = "README.md"
requires-python = ">=3.11"

[dependency-groups]
dev = [
    "build",
    "copier",
    "ipython",
    "import-linter",
    "myst-parser",
    "pre-commit",
    "pydata-sphinx-theme>=0.12",
    # Pin pyright until https://github.com/microsoft/pyright/issues/11060 is fixed
    "pyright==1.1.406",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-random-order",
    "ruff",
    "sphinx-autobuild",
    "sphinx-copybutton",
    "sphinx-design",
    "tox-uv",
    "types-mock",
    "types-requests",
    "softioc",
]

[project.scripts]
i19-bluesky = "i19_bluesky.__main__:main"

[project.urls]
GitHub = "https://github.com/DiamondLightSource/i19-bluesky"

[[project.authors]] # Further authors may be added by duplicating this section
email = "noemi.frisina@diamond.ac.uk"
name = "Noemi Frisina"

[tool.setuptools_scm]
version_file = "src/i19_bluesky/_version.py"

[tool.pyright]
typeCheckingMode = "standard"
reportMissingImports = false  # Ignore missing stubs in imported modules

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "skip_log_setup: marks tests so that loggers are not setup before the test.",
    "system_test: marks test as system test that requires additional infrastructure",
]
# Run pytest with all our checkers, and don't spam us with massive tracebacks on error
addopts = """
    --tb=native -vv --doctest-modules --doctest-glob="*.rst"
    """
# https://iscinumpy.gitlab.io/post/bound-version-constraints/#watch-for-warnings
filterwarnings = [
    "error",
    # Ignore deprecation warning from setuptools_dso (remove when https://github.com/epics-base/setuptools_dso/issues/36 is released)
    "ignore::DeprecationWarning:wheel",
]
# Doctest python code in docs, python code in src docstrings, test functions in tests
testpaths = "docs src tests"

[tool.coverage.run]
patch = ["subprocess"]
data_file = "/tmp/i19_bluesky.coverage"

[tool.coverage.paths]
# Tests are run from installed location, map back to the src directory
source = ["src", "**/site-packages/"]

[tool.tox]
skipsdist = true
# envs to runs automatically with tox -p
env_list = ["pre-commit", "type-checking", "tests", "docs", "system-test"]

[tool.tox.env_run_base]
# Pass all environment variables to the test environments
passenv = ["*"]

[tool.tox.env.pre-commit]
description = "Run pre-commit on all files"
commands = [
    [
        "pre-commit",
        "run",
        "--all-files",
        "--show-diff-on-failure",
        { replace = "posargs", default = [
        ], extend = true },
    ],
]

[tool.tox.env.type-checking]
description = "Run pyright"
commands = [
    [
        "pyright",
        "--pythonpath",
        ".venv/bin/python",
        "src",
        "tests",
        { replace = "posargs", default = [
        ], extend = true },
    ],
]

[tool.tox.env.tests]
description = "Run tests with coverage"
commands = [
    [
        "pytest",
        "--cov=i19_bluesky",
        "--cov-report",
        "term",
        "--cov-report",
        "xml:cov.xml",
        "tests/unit_tests",
        { replace = "posargs", default = [
        ], extend = true },
    ],
]

[tool.tox.env.system-test]
description = "Run system tests"
commands = [
    [
        "pytest",
        "tests/system_tests",
        { replace = "posargs", default = [
        ], extend = true },
    ],
]

[tool.tox.env.docs]
description = "Run docs build with clean environment"
commands = [
    [
        "sphinx-build",
        "--fresh-env",
        "--show-traceback",
        "--fail-on-warning",
        "--keep-going",
        "docs",
        "build/html",
        { replace = "posargs", default = [
        ], extend = true },
    ],
]

[tool.tox.env.docs-autobuild]
description = "Run docs with autobuild on change"
commands = [
    [
        "sphinx-autobuild",
        "--show-traceback",
        "--watch",
        "README.md",
        "docs",
        "build/html",
        { replace = "posargs", default = [
        ], extend = true },
    ],
]

[tool.ruff]
src = ["src", "tests"]
line-length = 88
lint.select = [
    "B",   # flake8-bugbear - https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
    "C4",  # flake8-comprehensions - https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
    "E",   # pycodestyle errors - https://docs.astral.sh/ruff/rules/#error-e
    "F",   # pyflakes rules - https://docs.astral.sh/ruff/rules/#pyflakes-f
    "W",   # pycodestyle warnings - https://docs.astral.sh/ruff/rules/#warning-w
    "I",   # isort - https://docs.astral.sh/ruff/rules/#isort-i
    "UP",  # pyupgrade - https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "SLF", # self - https://docs.astral.sh/ruff/settings/#lintflake8-self
]
lint.extend-ignore = [
    "B008", # Allow inject("device")
]

[tool.ruff.lint.per-file-ignores]
# By default, private member access is allowed in tests
# See https://github.com/DiamondLightSource/python-copier-template/issues/154
# Remove this line to forbid private member access in tests
"tests/**/*" = ["SLF001"]


[tool.importlinter]
root_package = "i19_bluesky"

[[tool.importlinter.contracts]]
name = "EH modules do not import from each other"
type = "independence"
modules = ["i19_bluesky.eh1", "i19_bluesky.eh2"]

[[tool.importlinter.contracts]]
name = "EH modules do not import from the optics module"
type = "forbidden"
source_modules = ["i19_bluesky.eh1", "i19_bluesky.eh2"]
forbidden_modules = "i19_bluesky.optics"

[[tool.importlinter.contracts]]
name = "Common plans module does not import from apps"
type = "forbidden"
source_modules = "i19_bluesky.plans"
forbidden_modules = [
    "i19_bluesky.eh1",
    "i19_bluesky.eh2",
    "i19_bluesky.serial",
    "i19_bluesky.optics",
]

[[tool.importlinter.contracts]]
name = "Common modules do not import from the optics module"
type = "forbidden"
source_modules = ["i19_bluesky.serial", "i19_bluesky.plans"]
forbidden_modules = "i19_bluesky.optics"
